add static ip validation

+++ DNSNetworkInterfacesManager.psm1
@@ -1,6 +1,5 @@
 using module .\DNSNetworkInterface.psm1
 class DNSNetworkInterfacesManager {
-    
     DNSNetworkInterfaces() {
     }
     [System.Collections.ArrayList]GetDNSNetworkInterfaces(){
@@ -22,13 +21,19 @@ class DNSNetworkInterfacesManager {
         return $networkInterfacesList
     }
 
-    SetStaticIPDNS([string]$interfaceIndex){
+    [string]SetStaticIPDNS([string]$interfaceIndex){
         $interfaceIpConfiguration =  Get-NetIPConfiguration -InterfaceIndex $interfaceIndex 
-        $interfaceDNSConfiguration = Get-DnsClientServerAddress -InterfaceIndex $interfaceIndex  -AddressFamily IPv4 
-        Remove-NetRoute -InterfaceIndex  $interfaceIndex -AddressFamily IPv4 -Confirm:$false
-        Remove-NetIPAddress -InterfaceIndex  $interfaceIndex  -AddressFamily IPv4  -Confirm:$false
-        Set-DnsClientServerAddress -InterfaceIndex $interfaceIndex -ServerAddresses ($interfaceIpConfiguration.IPv4Address.IPAddress, $interfaceDNSConfiguration.ServerAddresses[0]) 
-        New-NetIpAddress -InterfaceIndex $interfaceIndex  -IPAddress $interfaceIpConfiguration.IPv4Address.IPAddress -PrefixLength  $interfaceIpConfiguration.IPv4Address.PrefixLength -DefaultGateway $interfaceIpConfiguration.IPv4DefaultGateway.NextHop  -AddressFamily IPv4
+        if ($interfaceIpConfiguration.IPv4Address.PrefixOrigin -eq "Dhcp") {
+            $interfaceDNSConfiguration = Get-DnsClientServerAddress -InterfaceIndex $interfaceIndex  -AddressFamily IPv4 
+            Remove-NetRoute -InterfaceIndex  $interfaceIndex -AddressFamily IPv4 -Confirm:$false
+            Remove-NetIPAddress -InterfaceIndex  $interfaceIndex  -AddressFamily IPv4  -Confirm:$false
+            Set-DnsClientServerAddress -InterfaceIndex $interfaceIndex -ServerAddresses ($interfaceIpConfiguration.IPv4Address.IPAddress, $interfaceDNSConfiguration.ServerAddresses[0]) 
+            New-NetIpAddress -InterfaceIndex $interfaceIndex  -IPAddress $interfaceIpConfiguration.IPv4Address.IPAddress -PrefixLength  $interfaceIpConfiguration.IPv4Address.PrefixLength -DefaultGateway $interfaceIpConfiguration.IPv4DefaultGateway.NextHop  -AddressFamily IPv4
+            return "the ethernet interface is configured as static IP"
+        }
+        else {
+            return "the ethernet interface have already an static IP"
+        }
     }
 }
